공통 하단 Nav바가 추가됨에 따라 기존 햄버거바 정리 필요 : refactor : nav 폴더 정리, CommonNav.tsx common 폴더로 이관 https://github.com/Sejong-Balsamic/Malsami-FE/issues/451
