+++ src/deprecated/DDetailNav.tsx
