+++ src/deprecated/MyPageDetailNav.tsx
