+++ src/deprecated/QDetailNav.tsx
